misc: core/eme
